+++ Packages/Granita/Client/Worktrees/Domain/WorktreeReadTrigger.swift
@@ -1,5 +1,6 @@
 public enum WorktreeReadTrigger: Hashable, Sendable {
     case appearance
+    case returnToForeground
     case pullToRefresh
     case retry
 }